% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{test}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AA}{\code{\link[AA]{test}}}
}}
\keyword{internal}

